aider: Fixed closing brace and PHP tag placement in the `activate` method and removed extra closing brace at the end of the file.
# Aider chat conversation:

USER: # Fix any errors below, if possible.

## See relevant line below marked with █.

includes/class-rapid-url-indexer-activator.php:
...⋮...
  2│class Rapid_URL_Indexer_Activator {
  3│    public static function activate() {
  4│        global $wpdb;
  5│
  6│        $charset_collate = $wpdb->get_charset_collate();
  7│
  8│        // Create credits table
  9│        $table_name = $wpdb->prefix . 'rapid_url_indexer_credits';
 10│        $sql = "CREATE TABLE $table_name (
 11│            id mediumint(9) NOT NULL AUTO_INCREMENT,
...⋮...
 46│            wp_schedule_event(time(), 'hourly', 'rui_cron_job');
 47│    }
 48│}
 49█?>

ASSISTANT: To fix the error in the provided code, we need to ensure that the `activate` method is properly closed and the PHP closing tag is correctly placed. The issue seems to be with the missing closing brace for the `if` statement and the incorrect placement of the PHP closing tag.

Here is the corrected code:

includes/class-rapid-url-indexer-activator.php
```php
<<<<<<< SEARCH
            wp_schedule_event(time(), 'hourly', 'rui_cron_job');
=======
            wp_schedule_event(time(), 'hourly', 'rui_cron_job');
        }
>>>>>>> REPLACE
```

includes/class-rapid-url-indexer-activator.php
```php
<<<<<<< SEARCH
    }
=======
    }
}
>>>>>>> REPLACE
```

includes/class-rapid-url-indexer-activator.php
```php
<<<<<<< SEARCH
?>
=======
>>>>>>> REPLACE
```

USER:
# Fix any errors below, if possible.

## See relevant line below marked with █.

includes/class-rapid-url-indexer-activator.php:
...⋮...
 47│        }
 48│    }
 49│}
 50█}

ASSISTANT: The error in the provided code is due to an extra closing brace at the end of the file. We need to remove the extra closing brace to fix the issue.

Here is the corrected code:

includes/class-rapid-url-indexer-activator.php
```php
<<<<<<< SEARCH
    }
}
}
=======
    }
}
>>>>>>> REPLACE
```