aider: Added a custom field for the credits amount to simple products in WooCommerce.
# Aider chat conversation:

USER: when i create a simple product in woocommerce, there is no input field for the credits amount that will be added to the customers account after successfully purchasing the product

ASSISTANT: To add a custom field for the credits amount to simple products in WooCommerce, you can use the `woocommerce_product_options_general_product_data` action hook to add the field to the product data metabox. Then, save the field value using the `woocommerce_process_product_meta` action hook.

Here are the changes to make in the `includes/class-rapid-url-indexer.php` file:

includes/class-rapid-url-indexer.php
```php
<<<<<<< SEARCH
    private static function define_hooks() {
        add_action('rui_cron_job', array('Rapid_URL_Indexer', 'process_cron_jobs')); // Hourly cron job to update project status
        add_action('rui_process_api_request', array('Rapid_URL_Indexer', 'process_api_request'), 10, 3);
        add_action('rest_api_init', array('Rapid_URL_Indexer', 'register_rest_routes'));
        add_action('wp_ajax_rui_search_logs', array('Rapid_URL_Indexer_Admin', 'ajax_search_logs'));
        add_action('wp_ajax_nopriv_rui_search_logs', array('Rapid_URL_Indexer_Admin', 'ajax_search_logs'));
    }
=======
    private static function define_hooks() {
        add_action('rui_cron_job', array('Rapid_URL_Indexer', 'process_cron_jobs')); // Hourly cron job to update project status
        add_action('rui_process_api_request', array('Rapid_URL_Indexer', 'process_api_request'), 10, 3);
        add_action('rest_api_init', array('Rapid_URL_Indexer', 'register_rest_routes'));
        add_action('wp_ajax_rui_search_logs', array('Rapid_URL_Indexer_Admin', 'ajax_search_logs'));
        add_action('wp_ajax_nopriv_rui_search_logs', array('Rapid_URL_Indexer_Admin', 'ajax_search_logs'));

        // Add credits amount field to simple product
        add_action('woocommerce_product_options_general_product_data', array('Rapid_URL_Indexer', 'add_credits_field'));
        add_action('woocommerce_process_product_meta', array('Rapid_URL_Indexer', 'save_credits_field'));
    }

    public static function add_credits_field() {
        global $post;

        echo '<div class="options_group">';
        woocommerce_wp_text_input(array(
            'id' => '_credits_amount',
            'label' => __('Credits Amount', 'rapid-url-indexer'),
            'placeholder' => '',
            'desc_tip' => 'true',
            'description' => __('Enter the number of credits this product will add to the customer\'s account upon purchase.', 'rapid-url-indexer'),
            'type' => 'number',
            'custom_attributes' => array(
                'step' => '1',
                'min' => '0'
            )
        ));
        echo '</div>';
    }

    public static function save_credits_field($post_id) {
        $credits_amount = isset($_POST['_credits_amount']) ? intval($_POST['_credits_amount']) : 0;
        update_post_meta($post_id, '_credits_amount', $credits_amount);
    }
>>>>>>> REPLACE
```

This code adds a new "Credits Amount" field to the "General" tab of the simple product data metabox. The field allows entering a positive integer value representing the number of credits the product will add to the customer's account when purchased.

The `add_credits_field()` method outputs the field HTML, while the `save_credits_field()` method saves the entered value as post meta data when the product is saved.

With these changes, you'll be able to specify the credits amount for each simple product in WooCommerce. The plugin will then use this value to add credits to the customer's account when the product is purchased.