aider: The custom endpoints for "My Indexing Projects" and "Buy Indexing Credits" have been properly registered and their content has been set.
# Aider chat conversation:

USER: nothing changed, they still do not show up on the My Account page and also not as account endpoints in the woocommerce backend

ASSISTANT: The issue is that the `add_my_account_endpoints` method is not properly registering the endpoint for the "My Indexing Projects" page. It's missing the `add_action` call to set the content for that endpoint.

Here are the *SEARCH/REPLACE* blocks to fix it:

includes/class-rapid-url-indexer-customer.php
```php
<<<<<<< SEARCH
    public static function add_my_account_endpoints() {
        add_rewrite_endpoint('rui-buy-credits', EP_ROOT | EP_PAGES);
        add_rewrite_endpoint('rui-projects', EP_ROOT | EP_PAGES);
        add_action('woocommerce_account_rui-buy-credits_endpoint', array(__CLASS__, 'add_my_account_endpoint_content'));
    }
=======
    public static function add_my_account_endpoints() {
        add_rewrite_endpoint('rui-buy-credits', EP_ROOT | EP_PAGES);
        add_rewrite_endpoint('rui-projects', EP_ROOT | EP_PAGES);
        add_action('woocommerce_account_rui-buy-credits_endpoint', array(__CLASS__, 'add_my_account_endpoint_content'));
        add_action('woocommerce_account_rui-projects_endpoint', array(__CLASS__, 'projects_endpoint_content'));
    }
>>>>>>> REPLACE
```

This will properly register the endpoint for the "My Indexing Projects" page and set its content using the `projects_endpoint_content` method.

Also, to ensure the endpoints show up in the WooCommerce account endpoints list in the backend, we need to use the `woocommerce_account_menu_items` filter instead of `woocommerce_account_menu_items`:

includes/class-rapid-url-indexer-customer.php
```php
<<<<<<< SEARCH
    public static function add_my_account_menu_items($items) {
        $items['rui-projects'] = 'My Indexing Projects';
        $items['rui-buy-credits'] = 'Buy Indexing Credits';
        return $items;
    }
=======
    public static function add_my_account_menu_items($endpoints) {
        $endpoints['rui-projects'] = 'My Indexing Projects';
        $endpoints['rui-buy-credits'] = 'Buy Indexing Credits';
        return $endpoints;
    }
>>>>>>> REPLACE
```

And update the filter name in the `init` method:

includes/class-rapid-url-indexer-customer.php
```php
<<<<<<< SEARCH
        add_filter('woocommerce_account_menu_items', array(__CLASS__, 'add_my_account_menu_items'));
=======
        add_filter('woocommerce_account_menu_items', array(__CLASS__, 'add_my_account_menu_items'), 10, 1);
>>>>>>> REPLACE
```

With these changes, the custom endpoints should properly show up on the My Account page and in the WooCommerce account endpoints list in the backend.