aider: Changed product links to directly add them to the cart and forward the customer to the cart/checkout process.
# Aider chat conversation:

USER: on the rui-buy-credits page, can we change the links to the credits products to links that directly add them to the cart and forward the customer to the cart/checkout process when clicking one of the links?

ASSISTANT: To change the links to the credits products to directly add them to the cart and forward the customer to the cart/checkout process, we need to modify the `templates/customer-buy-credits.php` file.

Here is the *SEARCH/REPLACE* block:

templates/customer-buy-credits.php
```php
<<<<<<< SEARCH
                <a href="<?php echo esc_url($product->get_permalink()); ?>"><?php echo esc_html($product->get_name()); ?></a><span class="rui-credits-amount"> (adds <?php echo esc_html(get_post_meta($product->get_id(), '_credits_amount', true)); ?> credits to your account)</span>
=======
                <a href="<?php echo esc_url('?add-to-cart=' . $product->get_id()); ?>"><?php echo esc_html($product->get_name()); ?></a><span class="rui-credits-amount"> (adds <?php echo esc_html(get_post_meta($product->get_id(), '_credits_amount', true)); ?> credits to your account)</span>
>>>>>>> REPLACE
```

This change will make the product links add the product to the cart and redirect the customer to the cart page.