aider: Updated credits display to exclude the "Buy Credits" button on the "Remaining Credits" section of the "Buy Credits" page.
# Aider chat conversation:

USER: this adds the credits balance along with the buy credits button, right? the "remaining credits" on this page should not have the buy credits button

ASSISTANT: To ensure the "Remaining Credits" on the "Buy Credits" page does not include the "Buy Credits" button, we need to modify the `credits_display` function in `includes/class-rapid-url-indexer-customer.php` to accept a parameter that determines whether to show the "Buy Credits" button or not. Then, we will update the `templates/customer-buy-credits.php` to call this function without the button.

Here are the *SEARCH/REPLACE* blocks:

includes/class-rapid-url-indexer-customer.php
```php
<<<<<<< SEARCH
    public static function credits_display() {
=======
    public static function credits_display($show_button = true) {
>>>>>>> REPLACE
```

includes/class-rapid-url-indexer-customer.php
```php
<<<<<<< SEARCH
        return '<div class="rui-credits-display"><strong>Remaining Credits:</strong> ' . esc_html($credits) . ' <a href="' . esc_url(wc_get_endpoint_url('rui-buy-credits', '', wc_get_page_permalink('myaccount'))) . '" class="button wp-element-button">Buy Credits</a></div>';
=======
        $output = '<div class="rui-credits-display"><strong>Remaining Credits:</strong> ' . esc_html($credits) . '</div>';
        if ($show_button) {
            $output .= ' <a href="' . esc_url(wc_get_endpoint_url('rui-buy-credits', '', wc_get_page_permalink('myaccount'))) . '" class="button wp-element-button">Buy Credits</a>';
        }
        return $output;
>>>>>>> REPLACE
```

templates/customer-buy-credits.php
```php
<<<<<<< SEARCH
    <?php echo do_shortcode('[rui_credits_display]'); ?>
=======
    <?php echo Rapid_URL_Indexer_Customer::credits_display(false); ?>
>>>>>>> REPLACE
```

This will ensure that the "Remaining Credits" on the "Buy Credits" page does not include the "Buy Credits" button.