aider: Ensure "Buy Credits" button is displayed on "My Projects" page by updating the `templates/customer-projects.php` file to call the `credits_display` function with the `show_button` parameter set to `true`.
# Aider chat conversation:

USER: now the my projects page also does not show the buy credits button although it should

ASSISTANT: To ensure the "Buy Credits" button is displayed on the "My Projects" page, we need to update the `templates/customer-projects.php` file to call the `credits_display` function with the `show_button` parameter set to `true`.

Here is the *SEARCH/REPLACE* block:

templates/customer-projects.php
```php
<<<<<<< SEARCH
        <?php echo do_shortcode('[rui_credits_display]'); ?>
=======
        <?php echo Rapid_URL_Indexer_Customer::credits_display(true); ?>
>>>>>>> REPLACE
```

This will ensure that the "Buy Credits" button is displayed on the "My Projects" page.